requirements.txt
scapy  >= 2.4.5
PyYAML >= 6.0